Types

+++ Libs/Models/Contracts/ContractModel.cs
@@ -5,8 +5,8 @@
         public string id { get; set; } = string.Empty;
         public string name { get; set; } = string.Empty;
         public string description { get; set; } = string.Empty;
-        public decimal tickSize { get; set; }
-        public decimal tickValue { get; set; }
+        public double tickSize { get; set; }
+        public double tickValue { get; set; }
         public bool activeContract { get; set; }
         public string symbolId { get; set; } = string.Empty;
     }
