+++ Libs/Models/Gateway/GatewayTrade.cs
@@ -3,9 +3,9 @@
     public class GatewayTrade
     {
         public string symbolId { get; set; }
-        public decimal price { get; set; }
+        public double price { get; set; }
         public string timestamp { get; set; }
         public TradeLogType type { get; set; }
-        public decimal volume { get; set; }
+        public double volume { get; set; }
     }
 }
