+++ Libs/Models/Positions/PositionModel.cs
@@ -9,6 +9,6 @@
         public DateTime creationTimestamp { get; set; }
         public PositionType type { get; set; }
         public int size { get; set; }
-        public decimal averagePrice { get; set; }
+        public double averagePrice { get; set; }
     }
 }
